Move `ToFieldElements` into `to_field_elements.rs`
