+++ ledger/src/proofs/to_field_elements.rs
@@ -1,14 +1,35 @@
 use ark_ff::Field;
+use kimchi::proof::ProofEvaluations;
+use mina_curves::pasta::Fq;
 use mina_hasher::Fp;
-use mina_p2p_messages::v2;
+use mina_p2p_messages::{string::ByteString, v2};
+use mina_signer::CompressedPubKey;
 
-use crate::scan_state::{
-    fee_excess::FeeExcess,
-    scan_state::transaction_snark::{Registers, SokDigest, Statement},
-    transaction_logic::zkapp_statement::ZkappStatement,
+use crate::{
+    scan_state::{
+        currency::{self, Sgn},
+        fee_excess::FeeExcess,
+        pending_coinbase,
+        scan_state::transaction_snark::{Registers, SokDigest, Statement},
+        transaction_logic::{
+            protocol_state::{EpochData, EpochLedger},
+            transaction_union_payload,
+            zkapp_statement::ZkappStatement,
+        },
+    },
+    staged_ledger::hash::StagedLedgerHash,
+    Account, MyCow, ReceiptChainHash, TimingAsRecord, TokenId, TokenSymbol, VotingFor,
 };
 
-use super::witness::FieldWitness;
+use super::{
+    numbers::currency::{CheckedCurrency, CheckedSigned},
+    unfinalized::{AllEvals, EvalsWithPublicInput},
+    witness::{
+        field_to_bits, Boolean, FieldWitness, GroupAffine, InnerCurve, PlonkVerificationKeyEvals,
+        StepMainProofState, StepMainStatement,
+    },
+    wrap::CircuitVar,
+};
 
 pub trait ToFieldElements<F: Field> {
     fn to_field_elements(&self, fields: &mut Vec<F>);
@@ -214,5 +235,701 @@ impl<F: FieldWitness> ToFieldElements<F>
         fields.push(sign_to_field(&fee_excess_r.sgn));
 
         sok_digest.to_field_elements(fields)
+    }
+}
+
+impl<F: FieldWitness, T: ToFieldElements<F>> ToFieldElements<F> for Vec<T> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.iter().for_each(|v| v.to_field_elements(fields));
+    }
+}
+
+impl<F: FieldWitness, T: ToFieldElements<F>> ToFieldElements<F> for Box<[T]> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.iter().for_each(|v| v.to_field_elements(fields));
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for Fp {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        use crate::proofs::witness::IntoGeneric;
+        fields.push(self.into_gen());
+    }
+}
+
+// pack
+pub fn field_of_bits<F: FieldWitness, const N: usize>(bs: &[bool; N]) -> F {
+    bs.iter().rev().fold(F::zero(), |acc, b| {
+        let acc = acc + acc;
+        if *b {
+            acc + F::one()
+        } else {
+            acc
+        }
+    })
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for Fq {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        use crate::proofs::witness::IntoGeneric;
+        use std::any::TypeId;
+
+        // TODO: Refactor when specialization is stable
+        if TypeId::of::<F>() == TypeId::of::<Fq>() {
+            fields.push(self.into_gen());
+        } else {
+            // `Fq` is larger than `Fp` so we have to split the field (low & high bits)
+            // See:
+            // https://github.com/MinaProtocol/mina/blob/e85cf6969e42060f69d305fb63df9b8d7215d3d7/src/lib/pickles/impls.ml#L94C1-L105C45
+
+            let to_high_low = |fq: Fq| {
+                let [low, high @ ..] = field_to_bits::<Fq, 255>(fq);
+                [field_of_bits(&high), F::from(low)]
+            };
+            fields.extend(to_high_low(*self));
+        }
+    }
+}
+
+impl<F: FieldWitness, T: ToFieldElements<F>, const N: usize> ToFieldElements<F> for [T; N] {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.iter().for_each(|v| v.to_field_elements(fields));
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for StagedLedgerHash<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            non_snark,
+            pending_coinbase_hash,
+        } = self;
+
+        let non_snark_digest = non_snark.digest();
+
+        const BITS: [u8; 8] = [1, 2, 4, 8, 16, 32, 64, 128];
+        fields.extend(
+            non_snark_digest
+                .iter()
+                .flat_map(|byte| BITS.iter().map(|bit| F::from((*byte & bit != 0) as u64))),
+        );
+
+        pending_coinbase_hash.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for ByteString {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let slice: &[u8] = self;
+        slice.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for GroupAffine<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            x, y, infinity: _, ..
+        } = self;
+        y.to_field_elements(fields);
+        x.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for &'_ [u8] {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        const BITS: [u8; 8] = [1, 2, 4, 8, 16, 32, 64, 128];
+        fields.extend(
+            self.iter()
+                .flat_map(|byte| BITS.iter().map(|bit| F::from((*byte & bit != 0) as u64))),
+        );
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for &'_ [bool] {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        fields.reserve(self.len());
+        fields.extend(self.iter().copied().map(F::from))
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for bool {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        F::from(*self).to_field_elements(fields)
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for u64 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        F::from(*self).to_field_elements(fields)
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for u32 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        F::from(*self).to_field_elements(fields)
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for ProofEvaluations<[F; 2]> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            w,
+            z,
+            s,
+            coefficients,
+            generic_selector,
+            poseidon_selector,
+            complete_add_selector,
+            mul_selector,
+            emul_selector,
+            endomul_scalar_selector,
+            range_check0_selector,
+            range_check1_selector,
+            foreign_field_add_selector,
+            foreign_field_mul_selector,
+            xor_selector,
+            rot_selector,
+            lookup_aggregation,
+            lookup_table,
+            lookup_sorted,
+            runtime_lookup_table,
+            runtime_lookup_table_selector,
+            xor_lookup_selector,
+            lookup_gate_lookup_selector,
+            range_check_lookup_selector,
+            foreign_field_mul_lookup_selector,
+        } = self;
+
+        let mut push = |[a, b]: &[F; 2]| {
+            a.to_field_elements(fields);
+            b.to_field_elements(fields);
+        };
+
+        w.iter().for_each(&mut push);
+        coefficients.iter().for_each(&mut push);
+        push(z);
+        s.iter().for_each(&mut push);
+        push(generic_selector);
+        push(poseidon_selector);
+        push(complete_add_selector);
+        push(mul_selector);
+        push(emul_selector);
+        push(endomul_scalar_selector);
+        range_check0_selector.as_ref().map(&mut push);
+        range_check1_selector.as_ref().map(&mut push);
+        foreign_field_add_selector.as_ref().map(&mut push);
+        foreign_field_mul_selector.as_ref().map(&mut push);
+        xor_selector.as_ref().map(&mut push);
+        rot_selector.as_ref().map(&mut push);
+        lookup_aggregation.as_ref().map(&mut push);
+        lookup_table.as_ref().map(&mut push);
+        lookup_sorted.iter().for_each(|v| {
+            v.as_ref().map(&mut push);
+        });
+        runtime_lookup_table.as_ref().map(&mut push);
+        runtime_lookup_table_selector.as_ref().map(&mut push);
+        xor_lookup_selector.as_ref().map(&mut push);
+        lookup_gate_lookup_selector.as_ref().map(&mut push);
+        range_check_lookup_selector.as_ref().map(&mut push);
+        foreign_field_mul_lookup_selector.as_ref().map(&mut push);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for AllEvals<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            ft_eval1,
+            evals:
+                EvalsWithPublicInput {
+                    evals,
+                    public_input,
+                },
+        } = self;
+
+        public_input.to_field_elements(fields);
+        evals.to_field_elements(fields);
+        ft_eval1.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for &[AllEvals<F>] {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.iter().for_each(|e| e.to_field_elements(fields))
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for EpochData<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            ledger:
+                EpochLedger {
+                    hash,
+                    total_currency,
+                },
+            seed,
+            start_checkpoint,
+            lock_checkpoint,
+            epoch_length,
+        } = self;
+
+        hash.to_field_elements(fields);
+        total_currency.to_field_elements(fields);
+        seed.to_field_elements(fields);
+        start_checkpoint.to_field_elements(fields);
+        lock_checkpoint.to_field_elements(fields);
+        epoch_length.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::NonZeroCurvePoint {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let v2::NonZeroCurvePointUncompressedStableV1 { x, is_odd } = self.inner();
+
+        x.to_field::<F>().to_field_elements(fields);
+        is_odd.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F>
+    for v2::ConsensusProofOfStakeDataConsensusStateValueStableV2
+{
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let v2::ConsensusProofOfStakeDataConsensusStateValueStableV2 {
+            blockchain_length,
+            epoch_count,
+            min_window_density,
+            sub_window_densities,
+            last_vrf_output,
+            total_currency,
+            curr_global_slot:
+                v2::ConsensusGlobalSlotStableV1 {
+                    slot_number,
+                    slots_per_epoch,
+                },
+            global_slot_since_genesis,
+            staking_epoch_data,
+            next_epoch_data,
+            has_ancestor_in_same_checkpoint_window,
+            block_stake_winner,
+            block_creator,
+            coinbase_receiver,
+            supercharge_coinbase,
+        } = self;
+
+        let staking_epoch_data: EpochData<F> = staking_epoch_data.into();
+        let next_epoch_data: EpochData<F> = next_epoch_data.into();
+
+        blockchain_length.as_u32().to_field_elements(fields);
+        epoch_count.as_u32().to_field_elements(fields);
+        min_window_density.as_u32().to_field_elements(fields);
+        fields.extend(sub_window_densities.iter().map(|w| F::from(w.as_u32())));
+
+        {
+            let vrf: &[u8] = last_vrf_output.as_ref();
+            (&vrf[..31]).to_field_elements(fields);
+            // Ignore the last 3 bits
+            let last_byte = vrf[31];
+            for bit in [1, 2, 4, 8, 16] {
+                F::from(last_byte & bit != 0).to_field_elements(fields);
+            }
+        }
+
+        total_currency.as_u64().to_field_elements(fields);
+        slot_number.as_u32().to_field_elements(fields);
+        slots_per_epoch.as_u32().to_field_elements(fields);
+        global_slot_since_genesis.as_u32().to_field_elements(fields);
+        staking_epoch_data.to_field_elements(fields);
+        next_epoch_data.to_field_elements(fields);
+        has_ancestor_in_same_checkpoint_window.to_field_elements(fields);
+        block_stake_winner.to_field_elements(fields);
+        block_creator.to_field_elements(fields);
+        coinbase_receiver.to_field_elements(fields);
+        supercharge_coinbase.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::MinaBaseProtocolConstantsCheckedValueStableV1 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            k,
+            slots_per_epoch,
+            slots_per_sub_window,
+            delta,
+            genesis_state_timestamp,
+        } = self;
+
+        k.as_u32().to_field_elements(fields);
+        slots_per_epoch.as_u32().to_field_elements(fields);
+        slots_per_sub_window.as_u32().to_field_elements(fields);
+        delta.as_u32().to_field_elements(fields);
+        genesis_state_timestamp.as_u64().to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::MinaStateBlockchainStateValueStableV2 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            staged_ledger_hash,
+            genesis_ledger_hash,
+            ledger_proof_statement,
+            timestamp,
+            body_reference,
+        } = self;
+
+        let staged_ledger_hash: StagedLedgerHash<F> = staged_ledger_hash.into();
+
+        staged_ledger_hash.to_field_elements(fields);
+        genesis_ledger_hash
+            .inner()
+            .to_field::<F>()
+            .to_field_elements(fields);
+        ledger_proof_statement.to_field_elements(fields);
+        timestamp.as_u64().to_field_elements(fields);
+        body_reference.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::MinaStateProtocolStateBodyValueStableV2 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let v2::MinaStateProtocolStateBodyValueStableV2 {
+            genesis_state_hash,
+            blockchain_state,
+            consensus_state,
+            constants,
+        } = self;
+
+        genesis_state_hash
+            .inner()
+            .to_field::<F>()
+            .to_field_elements(fields);
+        blockchain_state.to_field_elements(fields);
+        consensus_state.to_field_elements(fields);
+        constants.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for TokenId {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self(token_id) = self;
+        token_id.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for CompressedPubKey {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self { x, is_odd } = self;
+        x.to_field_elements(fields);
+        is_odd.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for mina_signer::Signature {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self { rx, s } = self;
+
+        rx.to_field_elements(fields);
+        let s_bits = field_to_bits::<_, 255>(*s);
+        s_bits.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for mina_signer::PubKey {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let GroupAffine::<Fp> { x, y, .. } = self.point();
+        x.to_field_elements(fields);
+        y.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for transaction_union_payload::TransactionUnion {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        use transaction_union_payload::{Body, Common, TransactionUnionPayload};
+
+        let Self {
+            payload:
+                TransactionUnionPayload {
+                    common:
+                        Common {
+                            fee,
+                            fee_token,
+                            fee_payer_pk,
+                            nonce,
+                            valid_until,
+                            memo,
+                        },
+                    body:
+                        Body {
+                            tag,
+                            source_pk,
+                            receiver_pk,
+                            token_id,
+                            amount,
+                        },
+                },
+            signer,
+            signature,
+        } = self;
+
+        fee.to_field_elements(fields);
+        fee_token.to_field_elements(fields);
+        fee_payer_pk.to_field_elements(fields);
+        nonce.to_field_elements(fields);
+        valid_until.to_field_elements(fields);
+        memo.as_slice().to_field_elements(fields);
+        tag.to_untagged_bits().to_field_elements(fields);
+        source_pk.to_field_elements(fields);
+        receiver_pk.to_field_elements(fields);
+        token_id.to_field_elements(fields);
+        amount.to_field_elements(fields);
+        signer.to_field_elements(fields);
+        signature.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::MinaNumbersGlobalSlotSinceGenesisMStableV1 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.as_u32().to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for v2::MinaBasePendingCoinbaseStackVersionedStableV1 {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            data,
+            state: v2::MinaBasePendingCoinbaseStateStackStableV1 { init, curr },
+        } = self;
+
+        data.to_field::<F>().to_field_elements(fields);
+        init.to_field::<F>().to_field_elements(fields);
+        curr.to_field::<F>().to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for pending_coinbase::StateStack {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self { init, curr } = self;
+        init.to_field_elements(fields);
+        curr.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for pending_coinbase::Stack {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self {
+            data: pending_coinbase::CoinbaseStack(data),
+            state,
+        } = self;
+
+        data.to_field_elements(fields);
+        state.to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for TokenSymbol {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let field: Fp = self.to_field();
+        field.to_field_elements(fields);
+    }
+}
+
+// TODO: De-deduplicate with ToInputs
+impl<F: FieldWitness> ToFieldElements<F> for crate::Timing {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let TimingAsRecord {
+            is_timed,
+            initial_minimum_balance,
+            cliff_time,
+            cliff_amount,
+            vesting_period,
+            vesting_increment,
+        } = self.to_record();
+
+        F::from(is_timed).to_field_elements(fields);
+        F::from(initial_minimum_balance.as_u64()).to_field_elements(fields);
+        F::from(cliff_time.as_u32()).to_field_elements(fields);
+        F::from(cliff_amount.as_u64()).to_field_elements(fields);
+        F::from(vesting_period.as_u32()).to_field_elements(fields);
+        F::from(vesting_increment.as_u64()).to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for crate::Permissions<crate::AuthRequired> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.iter_as_bits(|bit| {
+            bit.to_field_elements(fields);
+        });
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for crate::AuthRequired {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        for bit in self.encode().to_bits() {
+            bit.to_field_elements(fields);
+        }
+    }
+}
+
+impl ToFieldElements<Fp> for Box<Account> {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Account {
+            public_key,
+            token_id: TokenId(token_id),
+            token_symbol,
+            balance,
+            nonce,
+            receipt_chain_hash: ReceiptChainHash(receipt_chain_hash),
+            delegate,
+            voting_for: VotingFor(voting_for),
+            timing,
+            permissions,
+            zkapp,
+        } = &**self;
+
+        // Important: Any changes here probably needs the same changes in `AccountUnhashed`
+        public_key.to_field_elements(fields);
+        token_id.to_field_elements(fields);
+        token_symbol.to_field_elements(fields);
+        balance.to_field_elements(fields);
+        nonce.to_field_elements(fields);
+        receipt_chain_hash.to_field_elements(fields);
+        let delegate = MyCow::borrow_or_else(delegate, CompressedPubKey::empty);
+        delegate.to_field_elements(fields);
+        voting_for.to_field_elements(fields);
+        timing.to_field_elements(fields);
+        permissions.to_field_elements(fields);
+        MyCow::borrow_or_default(zkapp)
+            .hash()
+            .to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for crate::MerklePath {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.hash().to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness, A: ToFieldElements<F>, B: ToFieldElements<F>> ToFieldElements<F> for (A, B) {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let (a, b) = self;
+        a.to_field_elements(fields);
+        b.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for ReceiptChainHash {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self(receipt_chain_hash) = self;
+        receipt_chain_hash.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for Sgn {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let field: F = self.to_field();
+        field.to_field_elements(fields)
+    }
+}
+
+impl<F: FieldWitness, T: currency::Magnitude + ToFieldElements<F>> ToFieldElements<F>
+    for currency::Signed<T>
+{
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self { magnitude, sgn } = self;
+
+        magnitude.to_field_elements(fields);
+        sgn.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for PlonkVerificationKeyEvals<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let Self {
+            sigma,
+            coefficients,
+            generic,
+            psm,
+            complete_add,
+            mul,
+            emul,
+            endomul_scalar,
+        } = self;
+
+        sigma.iter().for_each(|s| s.to_field_elements(fields));
+        coefficients
+            .iter()
+            .for_each(|c| c.to_field_elements(fields));
+        generic.to_field_elements(fields);
+        psm.to_field_elements(fields);
+        complete_add.to_field_elements(fields);
+        mul.to_field_elements(fields);
+        emul.to_field_elements(fields);
+        endomul_scalar.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness, const N: usize> ToFieldElements<F> for crate::address::raw::Address<N> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let zero = F::zero();
+        let one = F::one();
+
+        fields.extend(
+            self.iter()
+                .map(|b| match b {
+                    crate::Direction::Left => zero,
+                    crate::Direction::Right => one,
+                })
+                .rev(),
+        );
+    }
+}
+
+// Implementation for references
+impl<F: FieldWitness, T: ToFieldElements<F>> ToFieldElements<F> for &T {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        (*self).to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness, T: CheckedCurrency<F>> ToFieldElements<F> for CheckedSigned<F, T> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.sgn.to_field_elements(fields);
+        self.magnitude.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for InnerCurve<F> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        let GroupAffine::<F> { x, y, .. } = self.to_affine();
+        x.to_field_elements(fields);
+        y.to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for Boolean {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.to_field::<F>().to_field_elements(fields);
+    }
+}
+
+impl<F: FieldWitness> ToFieldElements<F> for CircuitVar<Boolean> {
+    fn to_field_elements(&self, fields: &mut Vec<F>) {
+        self.as_boolean().to_field_elements(fields);
+    }
+}
+
+impl ToFieldElements<Fp> for StepMainStatement {
+    fn to_field_elements(&self, fields: &mut Vec<Fp>) {
+        let Self {
+            proof_state:
+                StepMainProofState {
+                    unfinalized_proofs,
+                    messages_for_next_step_proof,
+                },
+            messages_for_next_wrap_proof,
+        } = self;
+
+        unfinalized_proofs.to_field_elements(fields);
+        messages_for_next_step_proof.to_field_elements(fields);
+        messages_for_next_wrap_proof.to_field_elements(fields);
     }
 }
